Split `MetricsUpdateTask` into multiple classes; Prefer JDOQL over SQL
