+++ src/main/java/org/dependencytrack/tasks/VulnerabilityAnalysisTask.java
@@ -23,9 +23,9 @@ import alpine.event.framework.Event;
 import alpine.event.framework.Subscriber;
 import org.apache.commons.collections4.CollectionUtils;
 import org.dependencytrack.event.InternalAnalysisEvent;
-import org.dependencytrack.event.MetricsUpdateEvent;
 import org.dependencytrack.event.OssIndexAnalysisEvent;
 import org.dependencytrack.event.PortfolioVulnerabilityAnalysisEvent;
+import org.dependencytrack.event.ProjectMetricsUpdateEvent;
 import org.dependencytrack.event.VulnDbAnalysisEvent;
 import org.dependencytrack.event.VulnerabilityAnalysisEvent;
 import org.dependencytrack.model.Component;
@@ -125,7 +125,7 @@ public class VulnerabilityAnalysisTask implements Subscriber {
         final PolicyEngine pe = new PolicyEngine();
         pe.evaluate(components);
         if (project != null) {
-            Event.dispatch(new MetricsUpdateEvent(project));
+            Event.dispatch(new ProjectMetricsUpdateEvent(project));
         }
     }
