+++ src/main/java/org/dependencytrack/tasks/metrics/ComponentMetricsUpdateTask.java
@@ -1,0 +1,254 @@
+/*
+ * This file is part of Dependency-Track.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) Steve Springett. All Rights Reserved.
+ */
+package org.dependencytrack.tasks.metrics;
+
+import alpine.common.logging.Logger;
+import alpine.event.framework.Event;
+import alpine.event.framework.Subscriber;
+import org.apache.commons.lang3.time.DurationFormatUtils;
+import org.dependencytrack.event.ComponentMetricsUpdateEvent;
+import org.dependencytrack.metrics.Metrics;
+import org.dependencytrack.model.Analysis;
+import org.dependencytrack.model.AnalysisState;
+import org.dependencytrack.model.Component;
+import org.dependencytrack.model.DependencyMetrics;
+import org.dependencytrack.model.Policy;
+import org.dependencytrack.model.PolicyViolation;
+import org.dependencytrack.model.Severity;
+import org.dependencytrack.model.ViolationAnalysis;
+import org.dependencytrack.model.ViolationAnalysisState;
+import org.dependencytrack.persistence.QueryManager;
+import org.dependencytrack.util.VulnerabilityUtil;
+
+import javax.jdo.PersistenceManager;
+import javax.jdo.Query;
+import javax.jdo.Transaction;
+import java.math.BigDecimal;
+import java.util.Date;
+import java.util.List;
+
+import static java.lang.Math.toIntExact;
+
+/**
+ * @since 4.6.0
+ */
+public class ComponentMetricsUpdateTask implements Subscriber {
+
+    private static final Logger LOGGER = Logger.getLogger(ComponentMetricsUpdateTask.class);
+
+    @Override
+    public void inform(final Event e) {
+        if (e instanceof final ComponentMetricsUpdateEvent event) {
+            try {
+                updateMetrics(event.component());
+            } catch (Exception ex) {
+                LOGGER.error("An unexpected error occurred while updating component metrics");
+            }
+        }
+    }
+
+    static Counters updateMetrics(Component component) throws Exception {
+        LOGGER.debug("Executing metrics update for component " + component.getUuid());
+        final var counters = new Counters();
+
+        try (final var qm = new QueryManager()) {
+            final PersistenceManager pm = qm.getPersistenceManager();
+
+            component = pm.getObjectById(Component.class, component.getId());
+            if (component == null) {
+                LOGGER.error("");
+                return null;
+            }
+
+            for (final VulnerabilityProjection vulnerability : getVulnerabilities(pm, component)) {
+                counters.vulnerabilities++;
+
+                // Replicate the behavior of Vulnerability#getSeverity
+                final Severity severity;
+                if (vulnerability.severity != null) {
+                    severity = Severity.valueOf(vulnerability.severity);
+                } else {
+                    severity = VulnerabilityUtil.getSeverity(vulnerability.cvssV2BaseScore, vulnerability.cvssV3BaseScore);
+                }
+
+                switch (severity) {
+                    case CRITICAL -> counters.critical++;
+                    case HIGH -> counters.high++;
+                    case MEDIUM -> counters.medium++;
+                    case LOW, INFO -> counters.low++;
+                    case UNASSIGNED -> counters.unassigned++;
+                }
+            }
+
+            counters.findingsTotal = toIntExact(counters.vulnerabilities);
+            counters.findingsAudited = toIntExact(getTotalAuditedFindings(pm, component));
+            counters.findingsUnaudited = counters.findingsTotal - counters.findingsAudited;
+            counters.suppressions = toIntExact(getTotalSuppressedFindings(pm, component));
+            counters.inheritedRiskScore = Metrics.inheritedRiskScore(counters.critical, counters.high, counters.medium, counters.low, counters.unassigned);
+
+            for (final PolicyViolationProjection violation : getPolicyViolations(pm, component)) {
+                counters.policyViolationsTotal++;
+
+                switch (PolicyViolation.Type.valueOf(violation.type().name())) {
+                    case LICENSE -> counters.policyViolationsLicenseTotal++;
+                    case OPERATIONAL -> counters.policyViolationsOperationalTotal++;
+                    case SECURITY -> counters.policyViolationsSecurityTotal++;
+                }
+
+                switch (Policy.ViolationState.valueOf(violation.violationState().name())) {
+                    case FAIL -> counters.policyViolationsFail++;
+                    case WARN -> counters.policyViolationsWarn++;
+                    case INFO -> counters.policyViolationsInfo++;
+                }
+            }
+
+            if (counters.policyViolationsLicenseTotal > 0) {
+                counters.policyViolationsLicenseAudited = toIntExact(getTotalAuditedPolicyViolations(pm, component, PolicyViolation.Type.LICENSE));
+                counters.policyViolationsLicenseUnaudited = counters.policyViolationsLicenseTotal - counters.policyViolationsLicenseAudited;
+            }
+            if (counters.policyViolationsOperationalTotal > 0) {
+                counters.policyViolationsOperationalAudited = toIntExact(getTotalAuditedPolicyViolations(pm, component, PolicyViolation.Type.OPERATIONAL));
+                counters.policyViolationsOperationalUnaudited = counters.policyViolationsOperationalTotal - counters.policyViolationsOperationalAudited;
+            }
+            if (counters.policyViolationsSecurityTotal > 0) {
+                counters.policyViolationsSecurityAudited = toIntExact(getTotalAuditedPolicyViolations(pm, component, PolicyViolation.Type.SECURITY));
+                counters.policyViolationsSecurityUnaudited = counters.policyViolationsSecurityTotal - counters.policyViolationsSecurityAudited;
+            }
+
+            counters.policyViolationsAudited = counters.policyViolationsLicenseAudited +
+                    counters.policyViolationsOperationalAudited +
+                    counters.policyViolationsSecurityAudited;
+            counters.policyViolationsUnaudited = counters.policyViolationsTotal - counters.policyViolationsAudited;
+
+            Transaction trx = pm.currentTransaction();
+            try {
+                trx.begin();
+                final DependencyMetrics latestMetrics = qm.getMostRecentDependencyMetrics(component);
+                if (!counters.hasChanged(latestMetrics)) {
+                    LOGGER.debug("Metrics of component " + component.getUuid() + " did not change");
+                    latestMetrics.setLastOccurrence(counters.measuredAt);
+                } else {
+                    LOGGER.debug("Metrics of component " + component.getUuid() + " changed");
+                    final DependencyMetrics metrics = counters.createComponentMetrics(component);
+                    pm.makePersistent(metrics);
+                }
+                trx.commit();
+            } finally {
+                if (trx.isActive()) {
+                    trx.rollback();
+                }
+            }
+
+            if (component.getLastInheritedRiskScore() == null ||
+                    component.getLastInheritedRiskScore() != counters.inheritedRiskScore) {
+                LOGGER.debug("Updating inherited risk score of component " + component.getUuid());
+                trx = qm.getPersistenceManager().currentTransaction();
+                try {
+                    trx.begin();
+                    component.setLastInheritedRiskScore(counters.inheritedRiskScore);
+                    trx.commit();
+                } finally {
+                    if (trx.isActive()) {
+                        trx.rollback();
+                    }
+                }
+            }
+
+            LOGGER.debug("Completed metrics update for component " + component.getUuid() + " in " +
+                    DurationFormatUtils.formatDuration(new Date().getTime() - counters.measuredAt.getTime(), "mm:ss:SS"));
+        }
+
+        return counters;
+    }
+
+    private static List<VulnerabilityProjection> getVulnerabilities(final PersistenceManager pm, final Component component) throws Exception {
+        try (final Query<?> query = pm.newQuery(Query.SQL, """
+                SELECT
+                    "VULNERABILITY"."SEVERITY",
+                    "VULNERABILITY"."CVSSV2BASESCORE",
+                    "VULNERABILITY"."CVSSV3BASESCORE"
+                FROM "COMPONENTS_VULNERABILITIES"
+                    INNER JOIN "COMPONENT" ON "COMPONENT"."ID" = "COMPONENTS_VULNERABILITIES"."COMPONENT_ID"
+                    INNER JOIN "VULNERABILITY" ON "VULNERABILITY"."ID" = "COMPONENTS_VULNERABILITIES"."VULNERABILITY_ID"
+                    LEFT JOIN "ANALYSIS"
+                        ON "ANALYSIS"."COMPONENT_ID" = "COMPONENTS_VULNERABILITIES"."COMPONENT_ID"
+                        AND "ANALYSIS"."VULNERABILITY_ID" = "COMPONENTS_VULNERABILITIES"."VULNERABILITY_ID"
+                WHERE "COMPONENTS_VULNERABILITIES"."COMPONENT_ID" = ?
+                    AND ("ANALYSIS"."SUPPRESSED" IS NULL OR "ANALYSIS"."SUPPRESSED" = ?)
+                ORDER BY "VULNERABILITY"."ID"
+                """)) {
+            query.setParameters(component.getId(), false);
+            return List.copyOf(query.executeResultList(VulnerabilityProjection.class));
+        }
+    }
+
+    private static long getTotalAuditedFindings(final PersistenceManager pm, final Component component) throws Exception {
+        try (final Query<Analysis> query = pm.newQuery(Analysis.class)) {
+            query.setFilter("""
+                    component == :component &&
+                    suppressed == false &&
+                    analysisState != :notSet &&
+                    analysisState != :inTriage
+                    """);
+            query.setParameters(component, AnalysisState.NOT_SET, AnalysisState.IN_TRIAGE);
+            query.setResult("count(this)");
+            return query.executeResultUnique(Long.class);
+        }
+    }
+
+    private static long getTotalSuppressedFindings(final PersistenceManager pm, final Component component) throws Exception {
+        try (final Query<Analysis> query = pm.newQuery(Analysis.class)) {
+            query.setFilter("component == :component && suppressed == true");
+            query.setParameters(component);
+            query.setResult("count(this)");
+            return query.executeResultUnique(Long.class);
+        }
+    }
+
+    private static List<PolicyViolationProjection> getPolicyViolations(final PersistenceManager pm, final Component component) throws Exception {
+        try (final Query<PolicyViolation> query = pm.newQuery(PolicyViolation.class)) {
+            query.setFilter("component == :component && (analysis == null || analysis.suppressed == false)");
+            query.setParameters(component);
+            query.setResult("type, policyCondition.policy.violationState");
+            return List.copyOf(query.executeResultList(PolicyViolationProjection.class));
+        }
+    }
+
+    private static long getTotalAuditedPolicyViolations(final PersistenceManager pm, final Component component, final PolicyViolation.Type violationType) throws Exception {
+        try (final Query<ViolationAnalysis> query = pm.newQuery(ViolationAnalysis.class)) {
+            query.setFilter("""
+                    component == :component &&
+                    suppressed == false &&
+                    analysisState != :notSet &&
+                    policyViolation.type == :violationType
+                    """);
+            query.setParameters(component, ViolationAnalysisState.NOT_SET, violationType);
+            query.setResult("count(this)");
+            return query.executeResultUnique(Long.class);
+        }
+    }
+
+    public record PolicyViolationProjection(Enum<?> type, Enum<?> violationState) {
+
+    }
+
+    public record VulnerabilityProjection(String severity, BigDecimal cvssV2BaseScore, BigDecimal cvssV3BaseScore) {
+    }
+
+}
