--- src/test/java/org/dependencytrack/tasks/MetricsUpdateTaskTest.java
@@ -1,776 +1,0 @@
-/*
- * This file is part of Dependency-Track.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * Copyright (c) Steve Springett. All Rights Reserved.
- */
-package org.dependencytrack.tasks;
-
-import alpine.event.framework.EventService;
-import org.dependencytrack.PersistenceCapableTest;
-import org.dependencytrack.event.CallbackEvent;
-import org.dependencytrack.event.MetricsUpdateEvent;
-import org.dependencytrack.model.AnalysisState;
-import org.dependencytrack.model.Component;
-import org.dependencytrack.model.DependencyMetrics;
-import org.dependencytrack.model.Policy;
-import org.dependencytrack.model.PolicyCondition;
-import org.dependencytrack.model.PolicyViolation;
-import org.dependencytrack.model.PortfolioMetrics;
-import org.dependencytrack.model.Project;
-import org.dependencytrack.model.ProjectMetrics;
-import org.dependencytrack.model.Severity;
-import org.dependencytrack.model.ViolationAnalysisState;
-import org.dependencytrack.model.Vulnerability;
-import org.dependencytrack.model.VulnerabilityMetrics;
-import org.dependencytrack.tasks.scanners.AnalyzerIdentity;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.Date;
-import java.util.List;
-import java.util.UUID;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class MetricsUpdateTaskTest extends PersistenceCapableTest {
-
-    @BeforeClass
-    public static void setUpClass() {
-        EventService.getInstance().subscribe(MetricsUpdateEvent.class, MetricsUpdateTask.class);
-        EventService.getInstance().subscribe(CallbackEvent.class, CallbackTask.class);
-    }
-
-    @AfterClass
-    public static void tearDownClass() {
-        EventService.getInstance().unsubscribe(MetricsUpdateTask.class);
-        EventService.getInstance().unsubscribe(CallbackTask.class);
-    }
-
-    @Test
-    public void testUpdatePortfolioMetricsEmpty() {
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
-
-        final PortfolioMetrics metrics = qm.getMostRecentPortfolioMetrics();
-        assertThat(metrics.getProjects()).isZero();
-        assertThat(metrics.getVulnerableProjects()).isZero();
-        assertThat(metrics.getComponents()).isZero();
-        assertThat(metrics.getVulnerableComponents()).isZero();
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-    }
-
-    @Test
-    public void testUpdatePortfolioMetricsUnchanged() {
-        // Record initial portfolio metrics
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
-        final PortfolioMetrics metrics = qm.getMostRecentPortfolioMetrics();
-        assertThat(metrics.getLastOccurrence()).isEqualTo(metrics.getFirstOccurrence());
-
-        // Run the task a second time, without any metric being changed
-        final var beforeSecondRun = new Date();
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
-
-        // Ensure that the lastOccurrence timestamp was correctly updated
-        qm.getPersistenceManager().refresh(metrics);
-        assertThat(metrics.getLastOccurrence()).isNotEqualTo(metrics.getFirstOccurrence());
-        assertThat(metrics.getLastOccurrence()).isAfterOrEqualTo(beforeSecondRun);
-    }
-
-    @Test
-    public void testUpdatePortfolioMetricsVulnerabilities() {
-        var vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-001");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.HIGH);
-        vuln = qm.createVulnerability(vuln, false);
-
-        // Create a project with an unaudited vulnerability.
-        var projectUnaudited = new Project();
-        projectUnaudited.setName("acme-app-a");
-        projectUnaudited = qm.createProject(projectUnaudited, List.of(), false);
-        var componentUnaudited = new Component();
-        componentUnaudited.setProject(projectUnaudited);
-        componentUnaudited.setName("acme-lib-a");
-        componentUnaudited = qm.createComponent(componentUnaudited, false);
-        qm.addVulnerability(vuln, componentUnaudited, AnalyzerIdentity.NONE);
-
-        // Create a project with an audited vulnerability.
-        var projectAudited = new Project();
-        projectAudited.setName("acme-app-b");
-        projectAudited = qm.createProject(projectAudited, List.of(), false);
-        var componentAudited = new Component();
-        componentAudited.setProject(projectAudited);
-        componentAudited.setName("acme-lib-b");
-        componentAudited = qm.createComponent(componentAudited, false);
-        qm.addVulnerability(vuln, componentAudited, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(componentAudited, vuln, AnalysisState.NOT_AFFECTED, null, null, null, false);
-
-        // Create a project with a suppressed vulnerability.
-        var projectSuppressed = new Project();
-        projectSuppressed.setName("acme-app-c");
-        projectSuppressed = qm.createProject(projectSuppressed, List.of(), false);
-        var componentSuppressed = new Component();
-        componentSuppressed.setProject(projectSuppressed);
-        componentSuppressed.setName("acme-lib-c");
-        componentSuppressed = qm.createComponent(componentSuppressed, false);
-        qm.addVulnerability(vuln, componentSuppressed, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(componentSuppressed, vuln, AnalysisState.FALSE_POSITIVE, null, null, null, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
-
-        final PortfolioMetrics metrics = qm.getMostRecentPortfolioMetrics();
-        assertThat(metrics.getProjects()).isEqualTo(3);
-        assertThat(metrics.getVulnerableProjects()).isEqualTo(2); // Finding for one project is suppressed
-        assertThat(metrics.getComponents()).isEqualTo(3);
-        assertThat(metrics.getVulnerableComponents()).isEqualTo(2); // Finding for one component is suppressed
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getSuppressed()).isEqualTo(1);
-        assertThat(metrics.getFindingsTotal()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getFindingsAudited()).isEqualTo(1);
-        assertThat(metrics.getFindingsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getInheritedRiskScore()).isEqualTo(10.0);
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refreshAll(projectUnaudited, projectAudited, projectSuppressed,
-                componentUnaudited, componentAudited, componentSuppressed);
-        assertThat(projectUnaudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(projectAudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(projectSuppressed.getLastInheritedRiskScore()).isZero();
-        assertThat(componentUnaudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(componentAudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(componentSuppressed.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdatePortfolioMetricsPolicyViolations() {
-        // Create a project with an unaudited violation.
-        var projectUnaudited = new Project();
-        projectUnaudited.setName("acme-app-a");
-        projectUnaudited = qm.createProject(projectUnaudited, List.of(), false);
-        var componentUnaudited = new Component();
-        componentUnaudited.setProject(projectUnaudited);
-        componentUnaudited.setName("acme-lib-a");
-        componentUnaudited = qm.createComponent(componentUnaudited, false);
-        createPolicyViolation(componentUnaudited, Policy.ViolationState.FAIL, PolicyViolation.Type.LICENSE);
-
-        // Create a project with an audited violation.
-        var projectAudited = new Project();
-        projectAudited.setName("acme-app-b");
-        projectAudited = qm.createProject(projectAudited, List.of(), false);
-        var componentAudited = new Component();
-        componentAudited.setProject(projectAudited);
-        componentAudited.setName("acme-lib-b");
-        componentAudited = qm.createComponent(componentAudited, false);
-        final var violationAudited = createPolicyViolation(componentAudited, Policy.ViolationState.WARN, PolicyViolation.Type.OPERATIONAL);
-        qm.makeViolationAnalysis(componentAudited, violationAudited, ViolationAnalysisState.APPROVED, false);
-
-        // Create a project with a suppressed violation.
-        var projectSuppressed = new Project();
-        projectSuppressed.setName("acme-app-c");
-        projectSuppressed = qm.createProject(projectSuppressed, List.of(), false);
-        var componentSuppressed = new Component();
-        componentSuppressed.setProject(projectSuppressed);
-        componentSuppressed.setName("acme-lib-c");
-        componentSuppressed = qm.createComponent(componentSuppressed, false);
-        final var violationSuppressed = createPolicyViolation(componentSuppressed, Policy.ViolationState.INFO, PolicyViolation.Type.SECURITY);
-        qm.makeViolationAnalysis(componentSuppressed, violationSuppressed, ViolationAnalysisState.REJECTED, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.PORTFOLIO));
-
-        final PortfolioMetrics metrics = qm.getMostRecentPortfolioMetrics();
-        assertThat(metrics.getProjects()).isEqualTo(3);
-        assertThat(metrics.getVulnerableProjects()).isZero();
-        assertThat(metrics.getComponents()).isEqualTo(3);
-        assertThat(metrics.getVulnerableComponents()).isZero();
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsWarn()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsInfo()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsTotal()).isEqualTo(2);
-        assertThat(metrics.getPolicyViolationsAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refreshAll(projectUnaudited, projectAudited, projectSuppressed,
-                componentUnaudited, componentAudited, componentSuppressed);
-        assertThat(projectUnaudited.getLastInheritedRiskScore()).isZero();
-        assertThat(projectAudited.getLastInheritedRiskScore()).isZero();
-        assertThat(projectSuppressed.getLastInheritedRiskScore()).isZero();
-        assertThat(componentUnaudited.getLastInheritedRiskScore()).isZero();
-        assertThat(componentAudited.getLastInheritedRiskScore()).isZero();
-        assertThat(componentSuppressed.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateProjectMetricsEmpty() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(project)));
-
-        final ProjectMetrics metrics = qm.getMostRecentProjectMetrics(project);
-        assertThat(metrics.getComponents()).isZero();
-        assertThat(metrics.getVulnerableComponents()).isZero();
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refresh(project);
-        assertThat(project.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateProjectMetricsUnchanged() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        // Record initial project metrics
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(project)));
-        final ProjectMetrics metrics = qm.getMostRecentProjectMetrics(project);
-        assertThat(metrics.getLastOccurrence()).isEqualTo(metrics.getFirstOccurrence());
-
-        // Run the task a second time, without any metric being changed
-        final var beforeSecondRun = new Date();
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(project)));
-
-        // Ensure that the lastOccurrence timestamp was correctly updated
-        qm.getPersistenceManager().refresh(metrics);
-        assertThat(metrics.getLastOccurrence()).isNotEqualTo(metrics.getFirstOccurrence());
-        assertThat(metrics.getLastOccurrence()).isAfterOrEqualTo(beforeSecondRun);
-    }
-
-    @Test
-    public void testUpdateProjectMetricsVulnerabilities() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        var vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-001");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.HIGH);
-        vuln = qm.createVulnerability(vuln, false);
-
-        // Create a component with an unaudited vulnerability.
-        var componentUnaudited = new Component();
-        componentUnaudited.setProject(project);
-        componentUnaudited.setName("acme-lib-a");
-        componentUnaudited = qm.createComponent(componentUnaudited, false);
-        qm.addVulnerability(vuln, componentUnaudited, AnalyzerIdentity.NONE);
-
-        // Create a project with an audited vulnerability.
-        var componentAudited = new Component();
-        componentAudited.setProject(project);
-        componentAudited.setName("acme-lib-b");
-        componentAudited = qm.createComponent(componentAudited, false);
-        qm.addVulnerability(vuln, componentAudited, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(componentAudited, vuln, AnalysisState.NOT_AFFECTED, null, null, null, false);
-
-        // Create a project with a suppressed vulnerability.
-        var componentSuppressed = new Component();
-        componentSuppressed.setProject(project);
-        componentSuppressed.setName("acme-lib-c");
-        componentSuppressed = qm.createComponent(componentSuppressed, false);
-        qm.addVulnerability(vuln, componentSuppressed, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(componentSuppressed, vuln, AnalysisState.FALSE_POSITIVE, null, null, null, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(project)));
-
-        final ProjectMetrics metrics = qm.getMostRecentProjectMetrics(project);
-        assertThat(metrics.getComponents()).isEqualTo(3);
-        assertThat(metrics.getVulnerableComponents()).isEqualTo(2); // Finding for one component is suppressed
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getSuppressed()).isEqualTo(1);
-        assertThat(metrics.getFindingsTotal()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getFindingsAudited()).isEqualTo(1);
-        assertThat(metrics.getFindingsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getInheritedRiskScore()).isEqualTo(10.0);
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refreshAll(project, componentUnaudited, componentAudited, componentSuppressed);
-        assertThat(project.getLastInheritedRiskScore()).isEqualTo(10.0);
-        assertThat(componentUnaudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(componentAudited.getLastInheritedRiskScore()).isEqualTo(5.0);
-        assertThat(componentSuppressed.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateProjectMetricsPolicyViolations() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        // Create a component with an unaudited violation.
-        var componentUnaudited = new Component();
-        componentUnaudited.setProject(project);
-        componentUnaudited.setName("acme-lib-a");
-        componentUnaudited = qm.createComponent(componentUnaudited, false);
-        createPolicyViolation(componentUnaudited, Policy.ViolationState.FAIL, PolicyViolation.Type.LICENSE);
-
-        // Create a component with an audited violation.
-        var componentAudited = new Component();
-        componentAudited.setProject(project);
-        componentAudited.setName("acme-lib-b");
-        componentAudited = qm.createComponent(componentAudited, false);
-        final var violationAudited = createPolicyViolation(componentAudited, Policy.ViolationState.WARN, PolicyViolation.Type.OPERATIONAL);
-        qm.makeViolationAnalysis(componentAudited, violationAudited, ViolationAnalysisState.APPROVED, false);
-
-        // Create a component with a suppressed violation.
-        var componentSuppressed = new Component();
-        componentSuppressed.setProject(project);
-        componentSuppressed.setName("acme-lib-c");
-        componentSuppressed = qm.createComponent(componentSuppressed, false);
-        final var violationSuppressed = createPolicyViolation(componentSuppressed, Policy.ViolationState.INFO, PolicyViolation.Type.SECURITY);
-        qm.makeViolationAnalysis(componentSuppressed, violationSuppressed, ViolationAnalysisState.REJECTED, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(project)));
-
-        final ProjectMetrics metrics = qm.getMostRecentProjectMetrics(project);
-        assertThat(metrics.getComponents()).isEqualTo(3);
-        assertThat(metrics.getVulnerableComponents()).isZero();
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsWarn()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsInfo()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsTotal()).isEqualTo(2);
-        assertThat(metrics.getPolicyViolationsAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refreshAll(project, componentUnaudited, componentAudited, componentSuppressed);
-        assertThat(project.getLastInheritedRiskScore()).isZero();
-        assertThat(componentUnaudited.getLastInheritedRiskScore()).isZero();
-        assertThat(componentAudited.getLastInheritedRiskScore()).isZero();
-        assertThat(componentSuppressed.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateComponentMetricsEmpty() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        var component = new Component();
-        component.setProject(project);
-        component.setName("acme-lib");
-        component = qm.createComponent(component, false);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(component)));
-
-        final DependencyMetrics metrics = qm.getMostRecentDependencyMetrics(component);
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refresh(component);
-        assertThat(component.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateComponentMetricsUnchanged() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        var component = new Component();
-        component.setProject(project);
-        component.setName("acme-lib");
-        component = qm.createComponent(component, false);
-
-        // Record initial project metrics
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(component)));
-        final DependencyMetrics metrics = qm.getMostRecentDependencyMetrics(component);
-        assertThat(metrics.getLastOccurrence()).isEqualTo(metrics.getFirstOccurrence());
-
-        // Run the task a second time, without any metric being changed
-        final var beforeSecondRun = new Date();
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(component)));
-
-        // Ensure that the lastOccurrence timestamp was correctly updated
-        qm.getPersistenceManager().refresh(metrics);
-        assertThat(metrics.getLastOccurrence()).isNotEqualTo(metrics.getFirstOccurrence());
-        assertThat(metrics.getLastOccurrence()).isAfterOrEqualTo(beforeSecondRun);
-    }
-
-    @Test
-    public void testUpdateComponentMetricsVulnerabilities() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        var component = new Component();
-        component.setProject(project);
-        component.setName("acme-lib");
-        component = qm.createComponent(component, false);
-
-        // Create an unaudited vulnerability.
-        var vulnUnaudited = new Vulnerability();
-        vulnUnaudited.setVulnId("INTERNAL-001");
-        vulnUnaudited.setSource(Vulnerability.Source.INTERNAL);
-        vulnUnaudited.setSeverity(Severity.HIGH);
-        vulnUnaudited = qm.createVulnerability(vulnUnaudited, false);
-        qm.addVulnerability(vulnUnaudited, component, AnalyzerIdentity.NONE);
-
-        // Create an audited vulnerability.
-        var vulnAudited = new Vulnerability();
-        vulnAudited.setVulnId("INTERNAL-002");
-        vulnAudited.setSource(Vulnerability.Source.INTERNAL);
-        vulnAudited.setSeverity(Severity.MEDIUM);
-        vulnAudited = qm.createVulnerability(vulnAudited, false);
-        qm.addVulnerability(vulnAudited, component, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(component, vulnAudited, AnalysisState.NOT_AFFECTED, null, null, null, false);
-
-        // Create a suppressed vulnerability.
-        var vulnSuppressed = new Vulnerability();
-        vulnSuppressed.setVulnId("INTERNAL-003");
-        vulnSuppressed.setSource(Vulnerability.Source.INTERNAL);
-        vulnSuppressed.setSeverity(Severity.MEDIUM);
-        vulnSuppressed = qm.createVulnerability(vulnSuppressed, false);
-        qm.addVulnerability(vulnSuppressed, component, AnalyzerIdentity.NONE);
-        qm.makeAnalysis(component, vulnSuppressed, AnalysisState.FALSE_POSITIVE, null, null, null, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(component)));
-
-        final DependencyMetrics metrics = qm.getMostRecentDependencyMetrics(component);
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isEqualTo(1);
-        assertThat(metrics.getMedium()).isEqualTo(1); // One is suppressed
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getSuppressed()).isEqualTo(1);
-        assertThat(metrics.getFindingsTotal()).isEqualTo(2); // One is suppressed
-        assertThat(metrics.getFindingsAudited()).isEqualTo(1);
-        assertThat(metrics.getFindingsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getInheritedRiskScore()).isEqualTo(8.0);
-        assertThat(metrics.getPolicyViolationsFail()).isZero();
-        assertThat(metrics.getPolicyViolationsWarn()).isZero();
-        assertThat(metrics.getPolicyViolationsInfo()).isZero();
-        assertThat(metrics.getPolicyViolationsTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refresh(component);
-        assertThat(component.getLastInheritedRiskScore()).isEqualTo(8.0);
-    }
-
-    @Test
-    public void testUpdateComponentMetricsPolicyViolations() {
-        var project = new Project();
-        project.setName("acme-app");
-        project = qm.createProject(project, List.of(), false);
-
-        var component = new Component();
-        component.setProject(project);
-        component.setName("acme-lib");
-        component = qm.createComponent(component, false);
-
-        // Create an unaudited violation.
-        createPolicyViolation(component, Policy.ViolationState.FAIL, PolicyViolation.Type.LICENSE);
-
-        // Create an audited violation.
-        final PolicyViolation auditedViolation = createPolicyViolation(component, Policy.ViolationState.WARN, PolicyViolation.Type.OPERATIONAL);
-        qm.makeViolationAnalysis(component, auditedViolation, ViolationAnalysisState.APPROVED, false);
-
-        // Create a suppressed violation.
-        final PolicyViolation suppressedViolation = createPolicyViolation(component, Policy.ViolationState.INFO, PolicyViolation.Type.SECURITY);
-        qm.makeViolationAnalysis(component, suppressedViolation, ViolationAnalysisState.REJECTED, true);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(qm.getPersistenceManager().detachCopy(component)));
-
-        final DependencyMetrics metrics = qm.getMostRecentDependencyMetrics(component);
-        assertThat(metrics.getCritical()).isZero();
-        assertThat(metrics.getHigh()).isZero();
-        assertThat(metrics.getMedium()).isZero();
-        assertThat(metrics.getLow()).isZero();
-        assertThat(metrics.getUnassigned()).isZero();
-        assertThat(metrics.getVulnerabilities()).isZero();
-        assertThat(metrics.getSuppressed()).isZero();
-        assertThat(metrics.getFindingsTotal()).isZero();
-        assertThat(metrics.getFindingsAudited()).isZero();
-        assertThat(metrics.getFindingsUnaudited()).isZero();
-        assertThat(metrics.getInheritedRiskScore()).isZero();
-        assertThat(metrics.getPolicyViolationsFail()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsWarn()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsInfo()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsTotal()).isEqualTo(2);
-        assertThat(metrics.getPolicyViolationsAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsSecurityTotal()).isZero(); // Suppressed
-        assertThat(metrics.getPolicyViolationsSecurityAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsSecurityUnaudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsLicenseAudited()).isZero();
-        assertThat(metrics.getPolicyViolationsLicenseUnaudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalTotal()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalAudited()).isEqualTo(1);
-        assertThat(metrics.getPolicyViolationsOperationalUnaudited()).isZero();
-
-        qm.getPersistenceManager().refresh(component);
-        assertThat(component.getLastInheritedRiskScore()).isZero();
-    }
-
-    @Test
-    public void testUpdateVulnerabilityMetricsEmpty() {
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.VULNERABILITY));
-
-        final List<VulnerabilityMetrics> metrics = qm.getVulnerabilityMetrics();
-        assertThat(metrics).isEmpty();
-    }
-
-    @Test
-    public void testUpdateVulnerabilityMetrics() {
-        // Create vulnerability with published timestamp.
-        var vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-001");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.LOW);
-        vuln.setPublished(Date.from(LocalDateTime.of(2020, 10, 1, 6, 6, 6).toInstant(ZoneOffset.UTC)));
-        qm.createVulnerability(vuln, false);
-
-        // Create vulnerability with created timestamp.
-        vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-002");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.MEDIUM);
-        vuln.setCreated(Date.from(LocalDateTime.of(2021, 11, 1, 6, 6, 6).toInstant(ZoneOffset.UTC)));
-        qm.createVulnerability(vuln, false);
-
-        // Create vulnerability with created AND published timestamp.
-        vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-003");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.HIGH);
-        vuln.setCreated(Date.from(LocalDateTime.of(2022, 12, 1, 6, 6, 6).toInstant(ZoneOffset.UTC)));
-        vuln.setPublished(Date.from(LocalDateTime.of(2022, 12, 2, 6, 6, 6).toInstant(ZoneOffset.UTC)));
-        qm.createVulnerability(vuln, false);
-
-        // Create vulnerability without created or published timestamps.
-        vuln = new Vulnerability();
-        vuln.setVulnId("INTERNAL-004");
-        vuln.setSource(Vulnerability.Source.INTERNAL);
-        vuln.setSeverity(Severity.CRITICAL);
-        qm.createVulnerability(vuln, false);
-
-        new MetricsUpdateTask().inform(new MetricsUpdateEvent(MetricsUpdateEvent.Type.VULNERABILITY));
-
-        final List<VulnerabilityMetrics> metrics = qm.getVulnerabilityMetrics();
-        assertThat(metrics).hasSize(6);
-        assertThat(metrics).satisfiesExactlyInAnyOrder(
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2020);
-                    assertThat(vm.getMonth()).isNull();
-                    assertThat(vm.getCount()).isEqualTo(1);
-                },
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2020);
-                    assertThat(vm.getMonth()).isEqualTo(10);
-                    assertThat(vm.getCount()).isEqualTo(1);
-                },
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2021);
-                    assertThat(vm.getMonth()).isNull();
-                    assertThat(vm.getCount()).isEqualTo(1);
-                },
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2021);
-                    assertThat(vm.getMonth()).isEqualTo(11);
-                    assertThat(vm.getCount()).isEqualTo(1);
-                },
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2022);
-                    assertThat(vm.getMonth()).isNull();
-                    assertThat(vm.getCount()).isEqualTo(1);
-                },
-                vm -> {
-                    assertThat(vm.getYear()).isEqualTo(2022);
-                    assertThat(vm.getMonth()).isEqualTo(12);
-                    assertThat(vm.getCount()).isEqualTo(1);
-                }
-        );
-    }
-
-    private PolicyViolation createPolicyViolation(final Component component, final Policy.ViolationState violationState, final PolicyViolation.Type type) {
-        final var policy = qm.createPolicy(UUID.randomUUID().toString(), Policy.Operator.ALL, violationState);
-        final var policyCondition = qm.createPolicyCondition(policy, PolicyCondition.Subject.COORDINATES, PolicyCondition.Operator.MATCHES, "");
-        final var policyViolation = new PolicyViolation();
-
-        policyViolation.setComponent(component);
-        policyViolation.setPolicyCondition(policyCondition);
-        policyViolation.setTimestamp(new Date());
-        policyViolation.setType(type);
-        return qm.addPolicyViolationIfNotExist(policyViolation);
-    }
-
-}
